extract annotation check to a helper to avoid duplication with S9015
