+++ java-checks/src/main/java/org/sonar/java/checks/tests/MockFieldShouldUseMockAnnotationCheck.java
@@ -17,14 +17,10 @@
 package org.sonar.java.checks.tests;
 
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import org.sonar.check.Rule;
 import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
 import org.sonar.plugins.java.api.semantic.MethodMatchers;
-import org.sonar.plugins.java.api.semantic.Symbol;
-import org.sonar.plugins.java.api.semantic.SymbolMetadata;
 import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.ExpressionTree;
 import org.sonar.plugins.java.api.tree.MethodInvocationTree;
@@ -33,11 +29,6 @@ import org.sonar.plugins.java.api.tree.VariableTree;
 
 @Rule(key = "S9015")
 public class MockFieldShouldUseMockAnnotationCheck extends IssuableSubscriptionVisitor {
-
-  private static final String MOCKITO_EXTENSION = "org.mockito.junit.jupiter.MockitoExtension";
-  private static final String EXTEND_WITH_ANNOTATION = "org.junit.jupiter.api.extension.ExtendWith";
-  private static final String RUN_WITH_ANNOTATION = "org.junit.runner.RunWith";
-  private static final String MOCKITO_JUNIT_RUNNER_PREFIX = "org.mockito.junit.MockitoJUnitRunner";
 
   private static final MethodMatchers MOCK_METHOD = MethodMatchers.create()
     .ofTypes("org.mockito.Mockito")
@@ -55,7 +46,7 @@ public class MockFieldShouldUseMockAnnotationCheck extends IssuableSubscriptionV
   @Override
   public void visitNode(Tree tree) {
     ClassTree classTree = (ClassTree) tree;
-    if (!isMockitoManagedClass(classTree)) {
+    if (!MockitoManagedClassHelper.isMockitoManagedClass(classTree)) {
       return;
     }
     classTree.members().stream()
@@ -63,62 +54,6 @@ public class MockFieldShouldUseMockAnnotationCheck extends IssuableSubscriptionV
       .map(VariableTree.class::cast)
       .filter(MockFieldShouldUseMockAnnotationCheck::hasFieldMockInitializer)
       .forEach(field -> reportIssue(field.initializer(), MESSAGE));
-  }
-
-  private static boolean isMockitoManagedClass(ClassTree classTree) {
-    Symbol classSymbol = classTree.symbol();
-    return isAnnotatedWithMockitoExtension(classSymbol) || isAnnotatedWithMockitoJUnitRunner(classSymbol);
-  }
-
-  private static boolean isAnnotatedWithMockitoExtension(Symbol classSymbol) {
-    return checkMockitoExtensionInMetadata(classSymbol.metadata(), new HashSet<>());
-  }
-
-  private static boolean checkMockitoExtensionInMetadata(SymbolMetadata metadata, Set<Symbol> visited) {
-    List<SymbolMetadata.AnnotationValue> extendWithValues = metadata.valuesForAnnotation(EXTEND_WITH_ANNOTATION);
-    if (extendWithValues != null) {
-      for (SymbolMetadata.AnnotationValue av : extendWithValues) {
-        if (isMockitoExtensionClass(av.value())) {
-          return true;
-        }
-      }
-    }
-    for (SymbolMetadata.AnnotationInstance annotation : metadata.annotations()) {
-      Symbol annotationSymbol = annotation.symbol();
-      if (!visited.contains(annotationSymbol)) {
-        visited.add(annotationSymbol);
-        if (checkMockitoExtensionInMetadata(annotationSymbol.metadata(), visited)) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  private static boolean isMockitoExtensionClass(Object value) {
-    if (value instanceof Symbol symbol) {
-      return symbol.type().is(MOCKITO_EXTENSION);
-    }
-    if (value instanceof Object[] values) {
-      for (Object v : values) {
-        if (v instanceof Symbol symbol && symbol.type().is(MOCKITO_EXTENSION)) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  private static boolean isAnnotatedWithMockitoJUnitRunner(Symbol classSymbol) {
-    List<SymbolMetadata.AnnotationValue> runWithValues = classSymbol.metadata().valuesForAnnotation(RUN_WITH_ANNOTATION);
-    if (runWithValues != null && runWithValues.size() == 1) {
-      Object value = runWithValues.get(0).value();
-      if (value instanceof Symbol.TypeSymbol typeSymbol) {
-        String fqn = typeSymbol.type().fullyQualifiedName();
-        return fqn.equals(MOCKITO_JUNIT_RUNNER_PREFIX) || fqn.startsWith(MOCKITO_JUNIT_RUNNER_PREFIX + "$");
-      }
-    }
-    return false;
   }
 
   private static boolean hasFieldMockInitializer(VariableTree field) {
